+++ src/tm_devices/helpers/functions.py
@@ -178,8 +178,6 @@ def check_for_update(package_name: str = PACKAGE_NAME, index_name: str = "pypi")
         installed_version = importlib.metadata.version(package_name)
 
         # Get the version from the index
-        # This code mirrors code found in scripts/pypi_latest_version.py.
-        # If this code is updated, the script should be updated too.
         url = f"https://{index_name}.org/pypi/{package_name}/json"
         response = requests.get(url, timeout=10)
         releases = json.loads(response.text)["releases"]
@@ -302,6 +300,7 @@ def create_visa_connection(
     visa_library: str,
     *,
     retry_connection: bool = False,
+    second_connection_attempt_delay: int = 60,
 ) -> MessageBasedResource:
     """Create a VISA resource.
 
@@ -309,8 +308,10 @@ def create_visa_connection(
         device_config_entry: The device config entry.
         visa_library: A string containing the VISA library to use to create a ResourceManager.
         retry_connection: Boolean indicating if a second connection attempt should be made. If True,
-            two attempts are made to establish a VISA connection, with a 60-second delay in between
-            each attempt.
+            two attempts are made to establish a VISA connection, with a configurable delay in
+            between each attempt.
+        second_connection_attempt_delay: The number of seconds to wait in between the first and
+            second connection attempts when `retry_connection=True`.
 
     Returns:
         A VISA resource that can be passed into the device driver.
@@ -341,9 +342,27 @@ def create_visa_connection(
     # The broad except is because pyvisa_py can throw a base exception in the tcpip.py file
     except Exception as error_1:
         if not retry_connection:
-            message = f"Unable to establish a VISA connection to {resource_expression}"
-            raise ConnectionError(message) from error_1
-        time.sleep(60)  # wait 60 seconds and try again
+            error_message = (
+                f"Unable to establish a VISA connection to the "
+                f"{device_config_entry.device_type.value}"
+                f"{(' with the alias '+repr(device_config_entry.alias)) if device_config_entry.alias else ''}"  # noqa: E501
+                f" using the resource expression '{resource_expression}'"
+                f" and the {repr(visa_library) if visa_library else 'default'} VISA library"
+            )
+            _logger.error(error_message)  # noqa: TRY400
+            _logger.error(  # noqa: TRY400
+                "1st exception: %s.%s: %s",
+                error_1.__class__.__module__,
+                error_1.__class__.__qualname__,
+                error_1,
+            )
+            raise ConnectionError(error_message) from error_1
+        _logger.debug(
+            "Initial connection attempt failed, waiting %d second(s) "
+            "before re-attempting to create the VISA connection",
+            second_connection_attempt_delay,
+        )
+        time.sleep(second_connection_attempt_delay)
         try:
             # noinspection PyTypeChecker
             visa_object: MessageBasedResource = visa.ResourceManager(  # pyright: ignore[reportAssignmentType]
@@ -351,16 +370,42 @@ def create_visa_connection(
             ).open_resource(resource_expression)
         # The broad except is because pyvisa_py can throw a base exception in the tcpip.py file
         except Exception as error_2:
-            message = f"Unable to establish a VISA connection to {resource_expression}\n\n"
+            error_message = (
+                f"Unable to establish a VISA connection (after two tries) to the "
+                f"{device_config_entry.device_type.value}"
+                f"{(' with the alias '+repr(device_config_entry.alias)) if device_config_entry.alias else ''}"  # noqa: E501
+                f" using the resource expression '{resource_expression}'"
+                f" and the {repr(visa_library) if visa_library else 'default'} VISA library"
+            )
+            _logger.error(error_message)  # noqa: TRY400
+            _logger.error(  # noqa: TRY400
+                "1st exception: %s.%s: %s",
+                error_1.__class__.__module__,
+                error_1.__class__.__qualname__,
+                error_1,
+            )
+            _logger.error(  # noqa: TRY400
+                "2nd exception: %s.%s: %s:",
+                error_2.__class__.__module__,
+                error_2.__class__.__qualname__,
+                error_2,
+            )
+            ping_message = ""
+            ping_output = ""
             if device_config_entry.connection_type in {
                 ConnectionTypes.TCPIP,
                 ConnectionTypes.SOCKET,
             }:
-                message += (
-                    f"This is the current ping output for the device at "
-                    f"{device_config_entry.address}:\n{ping_address(device_config_entry.address)}"
+                ping_output = (
+                    ping_address(device_config_entry.address)
+                    or "\nno response returned or unable to find device\n"
                 )
-            raise ConnectionError(message) from error_2
+                ping_message = (
+                    f"\n\nThis is the current ping output for the device at "
+                    f"{device_config_entry.address}:\n"
+                )
+                _logger.debug("%s%s", ping_message.lstrip(), ping_output)
+            raise ConnectionError(error_message + ping_message + ping_output) from error_2
 
     return _configure_visa_object(visa_object, device_config_entry, visa_library)